+++ inst/examples/app.R
@@ -176,7 +176,11 @@ shinyApp(
         extra_elements_tab
       )
     ),
-    title = "shinyDashboardPlus"
+    title = "shinyDashboardPlus",
+    footer = dashboardFooter(
+      left_text = "By Divad Nojnarg",
+      right_text = "Zurich, 2019"
+    )
   ),
   server = function(input, output) {
     output$distPlot <- renderPlot({