+++ R/deps.R
@@ -11,15 +11,11 @@ appendDependencies <- function(x, value) {
 
 # Add dashboard dependencies to a tag object
 addDeps <- function(x, md) {
-  if (getOption("shiny.minified", TRUE)) {
-    adminLTE_js <- "app.min.js"
-    shinydashboard_js <- "shinydashboard.min.js"
-    adminLTE_css <- c("AdminLTE.min.css", "_all-skins.min.css")
-  } else {
-    adminLTE_js <- "app.js"
-    shinydashboard_js <- "shinydashboard.js"
-    adminLTE_css <- c("AdminLTE.css", "_all-skins.css")
-  }
+  
+  # always use minified files (https://www.minifier.org)
+  adminLTE_js <- "js/app.min.js"
+  shinydashboard_js <- "shinydashboard.js"
+  adminLTE_css <- c("css/AdminLTE.min.css", "css/_all-skins.min.css")
   
   # material design deps
   if (md) {
@@ -35,20 +31,13 @@ addDeps <- function(x, md) {
   }
   
   dashboardDeps <- list(
-    # vanilla adminLTE css from shinydashboard
-    htmltools::htmlDependency(
-      "AdminLTE", 
-      "2.0.6",
-      c(file = system.file("AdminLTE", package = "shinydashboard")),
-      #script = adminLTE_js,
-      stylesheet = adminLTE_css
-    ),
-    # custom js for shinydashboardPlus
+    # custom adminLTE js and css for shinydashboardPlus
     htmltools::htmlDependency(
       "shinydashboardPlus",
       as.character(utils::packageVersion("shinydashboardPlus")),
       c(file = system.file("shinydashboardPlus-0.6.0", package = "shinydashboardPlus")),
-      script = adminLTE_js
+      script = adminLTE_js,
+      stylesheet = adminLTE_css
     ),
     # shinydashboard css and js deps
     htmltools::htmlDependency(
@@ -63,7 +52,7 @@ addDeps <- function(x, md) {
       htmltools::htmlDependency(
         "materialDesign",
         as.character(utils::packageVersion("shinydashboardPlus")),
-        c(file = system.file("materialDesign", package = "shinydashboardPlus")),
+        c(file = system.file("materialDesign-1.0", package = "shinydashboardPlus")),
         script = c(md_js, ripples_js, md_init_js),
         stylesheet = c(md_bootstrap_css, ripples_css, md_adminLTE_css, md_skins_css)
       )