+++ NEWS.md
@@ -72,13 +72,13 @@ or redirect at the top of the page
 ## Breaking changes
 - change argument name in `dropdownItem()` (`boxPlus()`): "target" is replaced by "url".
 - Widely simplify the `rightSidebar()` function: remove `rightSidebarTabList()`, `rightSidebarTabItem()` and
-`rightSidebarPanel()` from the user interface. See [here](https://divadnojnarg.github.io/shinydashboardPlus/articles/rightSidebar.html) to discover how to set up 
+`rightSidebarPanel()` from the user interface. See [here](https://rinterface.github.io/shinydashboardPlus/articles/rightSidebar.html) to discover how to set up 
 a new `rightSidebar()`
 ## Major changes
 - add a "width" argument to the `rightSidebar()` (set to 230 pixels by default) to 
 improve customization. 
 - new function `setShadow()` to set shadow and hover effects on any elements
-- add an "active" argument to `rightSidebarTabContent()`: see [#4](https://github.com/DivadNojnarg/shinydashboardPlus/issues/4)
+- add an "active" argument to `rightSidebarTabContent()`: see [#4](https://github.com/RinteRface/shinydashboardPlus/issues/4)
 - add a "sidebar_background" argument to `dashboardPagePlus()`
 - new `verticalProgress()` bars!
 - new functions in the `rightSidebar()`: `rightSidebarMenu()`, `rightSidebarMenuItem()`, `menuIcon()`