+++ NEWS.md
@@ -1,3 +1,10 @@
+# shinydashboardPlus 0.7.5
+
+## Bug Fixes
+- Fix #36: make sure that a `carousel()` item is displayed when it is generated via a shiny Output function. 
+Thanks @daattali for the report
+
+
 # shinydashboardPlus 0.7.0
 ## Breaking Changes
 - remove all CSS functions, namely `setShadow()`, `setPulse()`, `setShake()` and